Add system-message file and update code review function (#7)
A new file, system-message.txt, has been added to hold and manage the system message. The 'code_review' function in ai_funcs.rs has been refactored to take its instructions from this file, instead of using a hard-coded string. This change simplifies the function and makes the system message easier to modify. A minor adjustment was also made to the error handling in main.rs.