Add LICENSE and README files (#9)
Added a LICENSE file stipulating the conditions under which the software can be used. Also included a README file detailing the installation and usage information of the code review bot. The README file is intended to serve as a guide for users during setup and regular use of the application.